% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{bertin_bandwidth}
\alias{bertin_bandwidth}
\title{Calculates bandwidth for smoothing}
\usage{
bertin_bandwidth(sigma, mu0, init_b, init_L, m)
}
\arguments{
\item{sigma}{Noise level of the curves. If unknown, can be estimated from
using \code{estimate_sigma}.}

\item{mu0}{Minimum density of time points. If unknown, can be estimated
using \code{estimate_density}.}

\item{init_b}{Initialised Hölder exponent.}

\item{init_L}{Initialised Hölder constant.}

\item{m}{Average number of sampling points per curve.}
}
\value{
A scalar or vector, depending on the inputs.
}
\description{
\code{bertin_bandwidth} calculates the bandwidth used for downstream smoothing.
Approach is detailed in the references.
}
\references{
Bertin L, (2004) - Minimax exact constant in sup-norm for
nonparametric regression with random design.
}
