% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{bertin_smoother}
\alias{bertin_smoother}
\title{Smooths curves using modified Nadaraya-Watson smoother}
\usage{
bertin_smoother(data, grid, bandwidth)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid}{Vector of sampling points which smoothing
is performed.}

\item{bandwidth}{Bandwidth used to smooth curves, usually
calculated from \code{bertin_bandwidth}.}
}
\value{
A matrix, with the rows representing curves
and columns the time points where smoothing is performed.
}
\description{
\code{bertin_smoother} performs curve smoothing using a kernel, referred to as
the Bertin kernel (see \code{bertin_kernel}). Usually used as an auxiliary
function for others such as \code{presmoothing}.
}
\references{
Bertin L, (2004) - Minimax exact constant in sup-norm for
nonparametric regression with random design.
}
