% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bias_kernel}
\alias{bias_kernel}
\title{Computes part of bias term involving kernels in adaptive estimation}
\usage{
bias_kernel(H, L, cst, h_grid)
}
\arguments{
\item{H}{Vector containing the Hölder exponents.}

\item{L}{Vector containing the Hölder constants.}

\item{cst}{Vector containing the constant term depending on the kernel.}

\item{h_grid}{Vector containing the grid of bandwidths.}
}
\value{
Vector with the same length as \code{H}, \code{L} and \code{cst}.
}
\description{
Computes part of bias term involving kernels in adaptive estimation
}
