% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{bw_cov}
\alias{bw_cov}
\title{Computes the optimal bandwidth for the covariance in terms of MISE}
\usage{
bw_cov(curves, params, h_grid, t_grid, inflate = TRUE)
}
\arguments{
\item{curves}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{params}{List, containing
\itemize{
\item \strong{$t} Vector of evaluation points on which the parameters were estimated on.
\item \strong{$H} Vector containing the estimated Hölder exponent.
\item \strong{$L} Vector containing the estimated Hölder constant.
\item \strong{$sigma} Vector containing the estimated noise.
\item \strong{$moments2} Vector containing the estimated variance of the curves.
\item \strong{$moments2prod} Matrix containing the variance of the product of curves.
}}

\item{h_grid}{Vector containing the grid of bandwidths.}

\item{t_grid}{Vector containing the evaluation points on which the curves will
be smoothed on.}

\item{inflate}{Boolean, indicating whether to inflate the bandwidths by
the discretization errors.}
}
\value{
Numeric, containing the optimal bandwidth.
}
\description{
Given a list of functional data, the optimal bandwidth for the covariance
function in terms of mean integrated squared error (MISE) is computed on
a grid of bandwidths. A plug-in bandwidth rule based on risk bounds is
computed.
}
