% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPCA.R
\name{bw_FPCA}
\alias{bw_FPCA}
\title{Compute fully adaptive FPCA bandwidth}
\usage{
bw_FPCA(
  data,
  param_smooth,
  h_grid,
  t_grid,
  psi_mat,
  lambda_vec,
  inflate = TRUE
)
}
\arguments{
\item{data}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{h_grid}{Vector of points for bandwidth optimization.}

\item{t_grid}{Vector of sampling points to smooth curves, should be in the
interval \verb{[0, 1]}.}

\item{psi_mat}{Matrix containing the eigenfunctions, with columns
representing the j-th eigenfunction and the rows are evaluated points on
a grid.}

\item{lambda_vec}{Vector of eigenvalues, whose length must match the number
of columns of \code{psi_mat}.}

\item{inflate}{Boolean, indicating whether to inflate bandwidth as a
correction for the discretization error from estimating regularity.}

\item{para_smooth}{List of estimated parameters.}
}
\description{
Performs estimation of the FPCA bandwidth for each index of the eigenvalues
and eigenfunctions separately, leading to 2 * J different bandwidths, where
J indicates the number of eigen-elements considered. Differences in bandwidths
are driven by the different constants specific to each eigen-element.
}
\references{
Wang S., Patilea V., Klutchnikoff N., (2023+) - Adaptive
Functional Principal Components Analysis
}
