% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{bw_mean}
\alias{bw_mean}
\title{Computes the optimal bandwidth for the mean function}
\usage{
bw_mean(curves, param_smooth, h_grid, t_grid)
}
\arguments{
\item{curves}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{h_grid}{Vector of points for bandwidth optimization.}

\item{t_grid}{Vector of sampling points to smooth curves.}

\item{params}{List of parameters, which can be obtained from
\code{estimate_parameters_FPCA}.}
}
\value{
Numeric, containing the optimal bandwidth for the mean function.
}
\description{
The optimal bandwidth in terms of mean integrated squared error (MISE)
for the purpose of mean function estimation is computed.
}
