% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{constant_d}
\alias{constant_d}
\title{Calculate constants associated to multi-fractional brownian motion.}
\usage{
constant_d(x, y)
}
\arguments{
\item{x}{Vector of Hurst index in one dimension.}

\item{y}{Vector of Hurst index in second dimension.}
}
\value{
A vector of length \code{x}/ \code{y} containing the constants.
}
\description{
Calculate constants associated to multi-fractional brownian motion.
}
