% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_diag}
\alias{correct_diag}
\title{Corrects the diagonal band of covariance function}
\usage{
correct_diag(mat_input, h)
}
\arguments{
\item{mat_input}{Matrix, representing the covariance function.}

\item{h}{Bandwidth used to smooth curve.}
}
\value{
Matrix, with corrected diagonal band.
}
\description{
Corrects the diagonal band of covariance function
}
\references{
Patilea V., Wang S. (2022+) - Adaptive Functional
Principal Components Analysis
}
