% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cov_emp}
\alias{cov_emp}
\title{Computes the empirical covariance function for functional data}
\usage{
cov_emp(curves_learn, curves_online)
}
\arguments{
\item{curves_learn}{List of learning curves, containing the following as elements:
\itemize{
\item \strong{$t} Vector of sampling points.
\item \strong{$x} Vector of observed points.
}}

\item{curves_online}{List of online curves, containing the following as elements:
\itemize{
\item \strong{t} Vector of sampling points.
\item \strong{$x} Vector of observed points.
}}
}
\value{
List, containing the following elements:
-\strong{$cov_learn} Matrix containing the covariance function from the learning set.
-\strong{$cov_online} List, containing the sampling points from the online set and
the interpolated covariance function.
}
\description{
Given a list of commonly sampled functional data, the empirical mean
is computed on the sampling points, before interpolation if an optional grid
of output is specified. Note that if the evaluation points are specified,
the covariance function is first constructed by centering each curve at
the observed points, before interpolation is done.
}
