% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPCA.R
\name{cov_FPCA}
\alias{cov_FPCA}
\title{Compute the FPCA covariance function}
\usage{
cov_FPCA(data, h, h_power, t_smooth, center = TRUE, sigma_true, diag_cor)
}
\arguments{
\item{data}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{h}{Numeric containing the optimal bandwidth.}

\item{h_power}{Numeric, power to raise the bandwidth to when estimating
sigma for the purposes of diagonal correction.}

\item{t_smooth}{Vector of sampling points to smooth curves. Covariance will
be computed at bi-dimensional grid points constructed from this vector.}

\item{center}{Boolean, indicating whether to center curves by their mean
function.}
}
\value{
Matrix of dimension \verb{length(t_smooth) x length(t_smooth)}.
}
\description{
Estimates the covariance function for the purposes of FPCA.
}
\references{
Wang S., Patilea V., Klutchnikoff N., (2023+) - Adaptive
Functional Principal Components Analysis
}
