% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{cov_ISE}
\alias{cov_ISE}
\title{Computes the covariance function based on MISE}
\usage{
cov_ISE(
  Y_list,
  xout,
  hout,
  param_list,
  h_power = 0.9,
  inflate_bw = TRUE,
  diag_cor = TRUE
)
}
\arguments{
\item{Y_list}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{xout}{Vector containing the evaluation points for smoothing curves.}

\item{hout}{Vector containing the grid of bandwidths to optimise over.}

\item{param_list}{List of estimated parameters, for example from the
function \code{estimate_parameters_FPCA}.}

\item{h_power}{Numeric, power to raise bandwidth to when estimating sigma
for the purposes of diagonal correction for the covariance function.}

\item{inflate_bw}{Boolean, indicating whether to inflate the bandwidths by
the discretization errors.}

\item{diag_cor}{Boolean, indicating whether to perform diagonal correction.}
}
\value{
Matrix containing the estimated covariance function.
}
\description{
Given a list of functional data, the covariance function is computed based on
a plug-in bandwidth rule which minimises the mean-integrated squared error (MISE).
An adaptive approach is taken by first estimating the regularity of the sample
paths, and then computing the risk bounds based on the estimated regularity.
}
