% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{cov_mercer}
\alias{cov_mercer}
\title{Computes the covariance function for some input eigenvalues
and eigenfunctions using Mercer's theorem.}
\usage{
cov_mercer(x_val, y_fun)
}
\arguments{
\item{x_val}{Vector of eigenvalues.}

\item{y_fun}{Matrix, with each column containing the j-th eigenfunction}
}
\value{
Matrix, containing the K x K covariance function, where K
is the number of rows of \code{y_fun}.
}
\description{
Computes the covariance function for some input eigenvalues
and eigenfunctions using Mercer's theorem.
}
