% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_optimised.R
\name{covariance_ll}
\alias{covariance_ll}
\title{Estimate covariance function using adaptive bandwidths}
\usage{
covariance_ll(
  curves,
  grid_bandwidth = lseq(0.005, 0.1, length.out = 151),
  grid_smooth = seq(0, 1, length.out = 101),
  k0 = 1,
  grid_param = seq(0.1, 0.9, length.out = 20),
  sigma = NULL,
  mu0 = NULL
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used in
estimation.}

\item{grid_param}{Grid of points to estimate parameters.}

\item{sigma}{Noise level, if known. Defaults to NULL, which estimates it.}

\item{mu0}{Density lower bound for time points. Defaults to NULL, which
estimates it.}
}
\value{
A list with two elements, containing
\itemize{
\item \strong{$Gamma} Estimated covariance function
\item \strong{$bw} Matrix of bandwidths used to smooth each curve.
}
}
\description{
\code{covariance_ll} estimates the covariance function using curves that
have been smoothed using an adaptive bandwidth.
}
\examples{
covariance_ll(curves = curves_list,
grid_bandwidth = seq(0, 1, length.out = 151),
grid_smooth = seq(0, 1, length.out = 101),
k0 = 1,
grid_param = seq(0, 1, length.out = 20),
sigma = 0.1, mu0 = 1, nvalues = 10)
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
