% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{covariance_mfbm}
\alias{covariance_mfbm}
\title{Generate covariance function associated to multi-fractional brownian motion.}
\usage{
covariance_mfbm(points, hurst, ...)
}
\arguments{
\item{points}{Vector of sampling points.}

\item{hurst}{Vector containing the Hurst function.}

\item{points_disto}{Vector containing the distorted time points.}

\item{norm}{TRUE/FALSE Whether to normalise the covariance by constants.}
}
\value{
A vector of length \code{x}/ \code{y} containing the constants.
}
\description{
Generate covariance function associated to multi-fractional brownian motion.
}
