% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_covariance_norm.R
\name{covariance_norm}
\alias{covariance_norm}
\title{Estimate covariance function using adaptive global bandwidth}
\usage{
covariance_norm(curves, grid_bandwidth, grid_smooth, k0, params)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used in
estimation.}

\item{params}{A tibble, containing the following parameters:
\itemize{
\item \strong{$t$} Sampling points.
\item \strong{$H} Estimated Hölder exponents.
\item \strong{$L} Estimated Hölder constants.
\item \strong{$sigma} Estimated noise.
\item \strong{$mu0} Estimated density of time points.
}}
}
\value{
A list containing
\itemize{
\item \strong{$cov} Estimated covariance function.
\item \strong{$bandwidth} Global bandwidth used to smooth each curve.
\item \strong{$constants} Tibble containing Hölder constants.
\item \strong{$kernel_int} Integrated kernel.
\item \strong{$Ngamma_ts} Harmonic mean of the number of points used to smooth
each curve at \code{t|s}.
\item \strong{$Ngamma_st} Transpose of \code{Ngamma_ts}.
\item \strong{$WN} Total number of points selected for estimation.
}
}
\description{
\code{covariance_norm} estimates the covariance function using curves that
have been smoothed using an adaptive global bandwidth.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
