% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{curves_select}
\alias{curves_select}
\title{Computes the weights for curve selection}
\usage{
curves_select(curves, x, h)
}
\arguments{
\item{curves}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{x}{Vector, containing the evaluation points.}

\item{h}{Vector or numeric, containing the bandwidth.}
}
\value{
Array of dimension \verb{G x N x H}, corresponding to the
number of evaluation points, curves, and bandwidth points respectively.
}
\description{
Given a list of functional data and a grid of points \code{x}, indicator functions
are computed which checks if there is at least one point along a curve
in the neighbourhood (defined by a bandwidth) of an evaluation point in \code{x}.
}
