% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_recursive.R
\name{cv_smooth}
\alias{cv_smooth}
\title{Performs presmoothing for the purpose of parameter estimation with
least squares cross-validation bandwidth}
\usage{
cv_smooth(curves, grid, n_sample)
}
\arguments{
\item{curves}{A list containing the raw data points with two entries:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Observed Points
}}

\item{grid}{Grid of points to estimate parameters. Must be a matrix, with
the rows representing the 3 time points, \verb{t1, t2, t3}, and the columns
are the points on the estimation grid.}

\item{n_sample}{Number of curves to use as learning set to learn the
cross-validated bandwidth. Uses uniform sampling without replacement.}
}
\value{
An \verb{G x N x 3} array, where \code{G} is the number of points on the
the grid and \code{N} is the number of curves.
}
\description{
\code{cv_smooth} presmooths raw curves, for the purpose of
parameter estimation. Uses least squares cross-validation to determine
a global bandwidth on a learning set, and using the median of learned
bandwidths to smooth across all curves.
}
