% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diagonal_bias}
\alias{diagonal_bias}
\title{Computes the diagonal bias of the covariance function}
\usage{
diagonal_bias(curves, t_grid, h, zeta, wm_list, W, WN)
}
\arguments{
\item{curves}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{t_grid}{Vector of evaluation points where curves are smoothed.}

\item{h}{Numeric containing the bandwidth (usually an optimal one).}

\item{zeta}{Numeric containing the power of the bandwidth.}

\item{wm_list}{List, where each element contains an Mi x G matrix of
Nadaraya-Watson weights, where Mi is the number of points along curve i and
G is the length of \code{t_grid}.}

\item{W}{Matrix of dimension G x N, where N is the number of curves. It
contains the weights for curves selection, resulting from the function \code{curves_select}.}

\item{WN}{Matrix of dimension G x G, containing the number of effective sample
size used for covariance function estimation.}
}
\value{
Matrix, containing the diagonal bias.
}
\description{
Computes the diagonal bias of the covariance function
}
