% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eigenfunctions.R
\name{efunctions_adaptive}
\alias{efunctions_adaptive}
\title{Estimate eigenfunctions of functional data with adaptive bandwidths}
\usage{
efunctions_adaptive(
  curves,
  grid_bandwidth,
  grid_smooth,
  k0,
  nfunctions = 10,
  params
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used in
estimation.}

\item{nfunctions}{The number of eigenfunctions to be kept.}

\item{params}{A tibble, containing the following parameters:
\itemize{
\item \strong{$t$} Sampling points.
\item \strong{$H} Estimated Hölder exponents.
\item \strong{$L} Estimated Hölder constants.
\item \strong{$sigma} Estimated noise.
\item \strong{$mu} Estimated density of time points.
estimates it.
}}
}
\value{
A list containing
\itemize{
\item Normalised eigenfunctions.
\item Bandwidth used for smoothing curves.
}
}
\description{
\code{efunctions_adaptive} estimates eigenfunctions using an adaptive bandwidth.
Curves can be irregularly sampled, possess heterogenous
degrees of smoothness and be relatively sparse (but should have
at least 20 points per curve). Uses the "smooth first, then estimate"
approach.
}
\references{
Patilea V., Wang S. (2022+) - Adaptive Functional
Principal Components Analysis
}
