% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eigen_interpolate}
\alias{eigen_interpolate}
\title{Perform FPCA by interpolation}
\usage{
eigen_interpolate(curves, grid_smooth, nelements)
}
\arguments{
\item{curves}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{grid_smooth}{Vector of sampling points to perform interpolation.}

\item{nelements}{Numeric containing the number of eigen-elements to return.}
}
\value{
List, with the following components
\itemize{
\item \strong{evalues} Vector, containing the normalised eigenvalues.
\item \strong{efunctions} Matrix, containing the normalised eigenfunctions.
}
}
\description{
Linearly interpolate curves and obtain eigen-elements by eigen-decomposition
of the empirical mean and covariance functions
}
