% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_optimised.R
\name{estimate_bandwidth_covariance}
\alias{estimate_bandwidth_covariance}
\title{Estimate adaptive bandwidths for covariance function}
\usage{
estimate_bandwidth_covariance(
  curves,
  grid_bandwidth,
  grid_smooth,
  k0,
  grid_param,
  sigma = NULL,
  mu0 = NULL
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used in
estimation.}

\item{grid_param}{Grid of points to estimate parameters.}

\item{sigma}{Noise level, if known. Defaults to NULL, which estimates it.}

\item{mu0}{Density lower bound for time points. Defaults to NULL, which
estimates it.}
}
\value{
A list with two elements, containing
\itemize{
\item A matrix of bandwidth for every point \verb{(t,s)}.
\item A list, containing
\itemize{
\item \strong{$moments} Estimated moments.
\item \strong{$constants} Estimated Hölder constants \code{H(t)} and \code{L(t)}.
\item \strong{$kernel-int} Integrated constant of kernel.
\item \strong{$N_gamma_ts} Harmonic mean of the number of points used for
smoothing each curve at \code{(t|s)}.
\item \strong{$N_gamma_st} Harmonic mean of the number of points used for
smoothing each curve at \code{(s|t)}.
\item \strong{$WN} Total number of curves selected for estimation.
}
}
}
\description{
\code{estimate_bandwidth_covariance} estimates the bandwidth at every point
\verb{(t,s)} that is used to smooth each curve for the purpose of
covariance function estimation.
}
\examples{
estimate_bandwidth_covariance(curves = curves_list,
grid_bandwidth = seq(0, 1, length.out = 151),
grid_smooth = seq(0, 1, length.out = 101),
k0 = 1,
grid_param = seq(0, 1, length.out = 20),
sigma = 0.1, mu0 = 1, nvalues = 10)
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
