% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eigenfunctions.R
\name{estimate_bandwidth_efunctions}
\alias{estimate_bandwidth_efunctions}
\title{Estimate adaptive bandwidths for eigenfunctions}
\usage{
estimate_bandwidth_efunctions(
  curves,
  grid_bandwidth,
  grid_smooth,
  k0,
  nfunctions,
  params
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed Points.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used in
estimation.}

\item{nfunctions}{The number of eigenfunctions to be kept.}

\item{params}{A tibble, containing the following parameters:
\itemize{
\item \strong{$t$} Sampling points.
\item \strong{$H} Estimated Hölder exponents.
\item \strong{$L} Estimated Hölder constants.
\item \strong{$sigma} Estimated noise.
\item \strong{$mu} Estimated density of time points.
}}
}
\value{
A vector of length nfunctions containing the adaptive bandwidth
for each eigenfunction.
}
\description{
\code{estimate_bandwidth_efunctions} returns a vector of bandwidths for each
eigenfunction, used for smoothing curves and is optimal for
eigenfunction estimation (in the \code{L2} norm).
}
