% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_optimised.R
\name{estimate_bandwidth_mean}
\alias{estimate_bandwidth_mean}
\title{Estimate adaptive bandwidth for mean estimation}
\usage{
estimate_bandwidth_mean(curves, params, grid_bandwidth, grid_smooth, k0)
}
\arguments{
\item{params}{Tibble of parameters, containing
\itemize{
\item \strong{$H} Estimated regularity.
\item \strong{$L} Estimated Hölder constant.
\item \strong{$sigma} Noise level of curves.
\item \strong{$mu0} Density lower bound for time points. Defaults to NULL,
which estimates it.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used
in estimation.}

\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}
}
\value{
A tibble, containing the input \code{params} and vector of
bandwidths.
}
\description{
\code{estimate_bandwidth_mean} estimates the adaptive bandwidth
used for mean function estimation.
}
