% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{estimate_density}
\alias{estimate_density}
\title{Estimate minimum density of sample points}
\usage{
estimate_density(data)
}
\arguments{
\item{data}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}
}
\value{
Numeric containing the minimum density.
}
\description{
Estimate the minimum density of sampling points using a kernel-density
estimator, leaving out the boundary points to account for boundary bias.
}
\references{
Wang S., Patilea V., Klutchnikoff N., (2023+) - Adaptive
Functional Principal Components Analysis
}
