% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_and_L_functions.R
\name{estimate_H0}
\alias{estimate_H0}
\title{Performs estimation of Hölder exponent}
\usage{
estimate_H0(data, center = FALSE)
}
\arguments{
\item{data}{A list, containing
\itemize{
\item \strong{$x} Smoothed data points.
}}

\item{center}{TRUE/FALSE, where TRUE induces an additional
centering in each term.}
}
\value{
A vector containing the estimated values at the
sampling points of presmoothed curves.
}
\description{
\code{estimate_H0} estimates the Hölder constant of presmoothed curves,
typically as output from the function \code{presmoothing}.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
