% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_H0_order}
\alias{estimate_H0_order}
\title{Estimates regularity at a point}
\usage{
estimate_H0_order(data, t0 = 0, k0 = 2, sigma)
}
\arguments{
\item{data}{A list containing the raw data points with two entries:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Observed Points
}}

\item{t0}{Time point at which \code{H(t0)} should be computed.}

\item{k0}{Number of closest neighbours to be used in estimation.
If \code{k0 < 2}, the regularity will be set to 1.}

\item{sigma}{Noise level of the curves, assumed to be known. If
unknown, one can use the \code{estimate_sigma} function to estimate it.}
}
\value{
Numeric for \code{H(t0)}.
}
\description{
Estimates the regularity of a curve at a fixed point \code{t0}.
Usually used as an auxiliary function for \code{estimate_H0_order_list} to
estimate the local regularity on a grid of points, and should be used
only for curves that have regularity less than 1.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2022) - Learning the
smoothness of noisy curves with application to online curve estimation.
}
