% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_holder_const}
\alias{estimate_holder_const}
\title{Performs estimation of parameters}
\usage{
estimate_holder_const(
  data,
  grid_estim = seq(0.2, 0.8, length.out = 20),
  sigma = NULL,
  mu0 = NULL,
  presmoothing = "splines"
)
}
\arguments{
\item{data}{A list containing the raw data points with two entries:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Observed Points
}}

\item{grid_estim}{Grid of points to estimate parameters.}

\item{sigma}{Noise level, if known. Defaults to NULL, which
estimates it.}

\item{mu0}{Density lower bound for time points. Defaults to NULL,
which estimates it.}

\item{presmoothing}{Presmoothing option. Can be either "splines",
"bertin" or FALSE. If FALSE, uses the order statistics approach instead
(see \code{estimate_H0_order}).}
}
\value{
Tibble with estimated parameters as columns.
}
\description{
\code{estimate_holder_const} estimates the parameters used for
downstream analysis, typically for functions such as
\code{evalues_adaptive}.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
