% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_recursive.R
\name{estimate_holder_quantities}
\alias{estimate_holder_quantities}
\title{Performs estimation of parameters}
\usage{
estimate_holder_quantities(
  curves,
  grid_param,
  weighted,
  cv,
  n_sample = 20,
  separate_curves,
  inflate_bw
)
}
\arguments{
\item{curves}{A list containing the raw data points with two entries:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Observed Points
}}

\item{grid_param}{Grid of points to estimate parameters.}

\item{weighted}{TRUE/FALSE, specifying if a weighted approach should be used,
which gives more weights to curves with more observed point around a fixed
\code{t}.}

\item{cv}{TRUE/FALSE, specifying whether cross validation should be used
on a learning set. If FALSE, uses an iterative plug-in local bandwidth
approach instead.}

\item{n_sample}{Integer. Specifies the number of curves to be used
to learn the cross-validation bandwidth.}

\item{separate_curves}{Boolean, indicating whether to use different curves for
presmoothing on different pairs of grid points \verb{(t1, t3)} and \verb{(t1, t2)}.}

\item{inflate_bw}{Boolean, indicating whether to inflate the bandwidth learned
from cross-validation by a constant corresponding log rate.}
}
\value{
A list with two elements:
\itemize{
\item \strong{$params} Tibble of parameters.
\item \strong{$smoothed_curves} Smoothed_curves.
}
}
\description{
\code{estimate_holder_quantities} estimates the parameters used for
downstream analysis, typically for functions such as
\code{evalues_adaptive}. Cross validation to determine the local
bandwidth for presmoothing, before a correction is performed
based on theoretical developments.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
