% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_recursive.R
\name{estimate_L}
\alias{estimate_L}
\title{Estimated the local Hölder constant based on presmoothing}
\usage{
estimate_L(smoothed_curves, t_list, H)
}
\arguments{
\item{smoothed_curves}{An array with dimensions \verb{G x N x 3}, where
the rows are the number of points on the grid used for estimation,
the columns are the curves, and the 3rd dimension represent the 3 time points.
Should use \code{locpoly_smooth} as an auxiliary function to obtain the proper
inputs.}

\item{t_list}{Must be a matrix, with
the rows representing the 3 time points, \verb{t1, t2, t3}, and the columns
are the points on the estimation grid.}

\item{H}{A vector with estimated regularity from \code{estimate_H}, with
the same length as the number of columns in \code{t_list}.}
}
\value{
A vector, whose length is equal to the number of sampling
points the presmoothed curves takes values in.
}
\description{
Estimates the local Hölder constant with supplied presmoothed curves.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
