% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_and_L_functions.R
\name{estimate_L0}
\alias{estimate_L0}
\title{Performs estimation of Hölder constant}
\usage{
estimate_L0(data, H0_list, M = NULL, center = FALSE)
}
\arguments{
\item{data}{A list, containing
\itemize{
\item \strong{$x} Smoothed data points.
}}

\item{H0_list}{Vector containing Hölder exponents, typically as
output from \code{estimate_H0}.}

\item{M}{Mean number of points on each curve. Defaults to NULL,
which calculates it.}

\item{center}{TRUE/FALSE, where TRUE induces an additional
centering in each term.}
}
\value{
A vector containing the estimated values at the
sampling points of presmoothed curves.
}
\description{
\code{estimate_L0} estimates the Hölder constant of presmoothed curves,
typically as output from the function \code{presmoothing}.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
