% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_L0_order}
\alias{estimate_L0_order}
\title{Estimates Hölder constant at a fixed point}
\usage{
estimate_L0_order(data, t0, H0, k0, sigma)
}
\arguments{
\item{data}{A list containing the raw data points with two entries:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Observed Points
}}

\item{t0}{Numeric, the sampling point to estimate at which \code{L(t0)}
should be estimated.}

\item{H0}{Numeric, the regularity \code{H(t0)}.}

\item{k0}{Number of neighbours to be used in estimation.}

\item{sigma}{Noise level of the curves, assumed to be known. If
unknown, one can use the \code{estimate_sigma} function to estimate it.}
}
\value{
A numeric.
}
\description{
Estimates the Hölder constant at a fixed point \code{t0}
using order statistics (see references). Should only be used for curves
with a regularity less than one.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2022) - Learning the
smoothness of noisy curves with application to online curve estimation.
}
