% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_L0_order_list}
\alias{estimate_L0_order_list}
\title{Estimates Hölder constant on a grid of points}
\usage{
estimate_L0_order_list(data, t0_list, k0_list, H0_list, sigma)
}
\arguments{
\item{data}{A list containing the raw data points with two entries:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Observed Points
}}

\item{t0_list}{Vector, containing the sampling points at which
\code{L(t)} should be estimated.}

\item{k0_list}{The number of neighbours to be considered in estimation
corresponding to the vector t0_list. To set the same number of
neighbours used in estimation, use a numeric.}

\item{H0_list}{Vector, containing the points \code{H(t)} estimated on the
grid \code{t0_list}. One can use \code{estimate_H0_order_list} to estimate \code{H(t)}
at \code{t0_list}.}

\item{sigma}{Noise level of the curves, assumed to be known. If
unknown, one can use the \code{estimate_sigma} function to estimate it.}
}
\value{
A vector of points.
}
\description{
Estimates the Hölder constant on a grid of points \code{t0_list}
using order statistics (see references). Makes use of the function
\code{estimate_L0_order}. Should only be used for curves
with a regularity less than one.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2022) - Learning the
smoothness of noisy curves with application to online curve estimation.
}
