% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{estimate_moments}
\alias{estimate_moments}
\title{Estimate moments for bandwidth constants}
\usage{
estimate_moments(presmoothed)
}
\arguments{
\item{presmoothed}{List, with each element containing vector of presmoothed
curves.}
}
\value{
A list, containing
\itemize{
\item \strong{$moment2} Second moment of \code{Xt}.
\item \strong{$moment2prod} Second moment of \code{XtXs}.
}
}
\description{
Performs estimation of the second moments of the curves \code{Xt} and products
\code{XtXs}. These moments appear as constants in bandwidth optimization problems.
Moments are estimated on the same grid as the presmoothed curves, and a
correction is made to adjust for sparse sampling schemes.
}
\references{
Wang S., Patilea V., Klutchnikoff N., (2023+) - Adaptive
Functional Principal Components Analysis
}
