% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{estimate_regularity}
\alias{estimate_regularity}
\title{Estimated the local regularity based on presmoothing}
\usage{
estimate_regularity(data, grid, bandwidth, intp = TRUE, gamma, nknots = 15)
}
\arguments{
\item{data}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{grid}{Vector of points to estimate the regularity.}

\item{bandwidth}{Bandwidth used to presmooth curves.}

\item{intp}{Boolean, indicating whether to use interpolation method. If
TRUE, presmoothing is first done at the observed time points with a
prespecified bandwidth, before interpolating back on the t1, t2, t3 grid.}

\item{gamma}{Numeric, power for computing delta.}

\item{nknots}{Numeric, number of knots to be used for smoothing H & L.}
}
\description{
Estimates the Hurst index \code{H} and Hölder constant \code{L} using presmoothing.
}
\references{
Wang S., Patilea V., Klutchnikoff N., (2023+) - Adaptive
Functional Principal Components Analysis
}
