% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{estimate_regularity_dense}
\alias{estimate_regularity_dense}
\title{Estimate the regularity of dense functional data, by only performing
interpolation and directly estimating H and L in the local neighbourhoods
defined by delta.}
\usage{
estimate_regularity_dense(data, xout, method, gamma_H, gamma_L)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{xout}{Vector of sampling points to estimate the regularity parameters.}

\item{method}{String, containing the interpolation method.}

\item{gamma_H}{Numeric, power to be used in computing delta for Hurst index.}

\item{gamma_L}{Numeric, power to be used in computinog delta for Holder
constant.}
}
\value{
List, containing the estimated H and L.
}
\description{
Estimate the regularity of dense functional data, by only performing
interpolation and directly estimating H and L in the local neighbourhoods
defined by delta.
}
