% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{estimate_regularity_deriv}
\alias{estimate_regularity_deriv}
\title{Estimate the regularity of differentiable functional data}
\usage{
estimate_regularity_deriv(data, grid, bandwidth, gamma, drv)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid}{Vector of sampling points to estimate the regularity parameters.}

\item{bandwidth}{Numeric, containing the bandwidth value.}

\item{gamma}{Numeric, power to be used in computing delta.}

\item{drv}{Numeric, containing the r-th order derivative to be used in estimation.}
}
\value{
List, containing the estimated H and L.
}
\description{
For differentiable sample paths, the derivatives are used to estimate the
local regularity parameters.
}
