% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_and_L_functions.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{Estimate noise level of curves}
\usage{
estimate_sigma(data)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}
}
\value{
A number.
}
\description{
Estimate noise level of curves
}
