% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_optimised.R
\name{estimate_variance_curves}
\alias{estimate_variance_curves}
\title{Estimate moments of curves for downstream analysis}
\usage{
estimate_variance_curves(data, params, grid_smooth, sigma = NULL, mu0 = NULL)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{params}{Tibble of parameters, containing
\itemize{
\item \strong{$H} Estimated regularity.
\item \strong{$L} Estimated Hölder constant.
}}

\item{grid_smooth}{Grid of points to smooth curves.}
}
\value{
A list, containing
\itemize{
\item \strong{$VarXt} Estimated Variance of \code{Xt}.
\item \strong{$VarXtXs} Estimated Variance of \code{XtXs}.
\item \strong{$EXt2} Estimated second moment of \code{Xt}.
\item \strong{$EXtXs2} Estimated second momenet of \code{XtXs}.
}
}
\description{
\code{estimate_variance_curves} estimates the moments used in
downstream adaptive estimation.
}
