% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{estimate_variance_irreg}
\alias{estimate_variance_irreg}
\title{Estimate the variance of raw data curves, which do not require
smoothing.}
\usage{
estimate_variance_irreg(data, grid_estim)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_estim}{Vector of sampling points to estimate the variance.}
}
\value{
Vector, comprising the estimated variance on the grid of
sampling points.
}
\description{
\code{estimate_variance_irreg} estimates the variance using only the
raw data points, possibly irregularly sampled, without pre-smoothing
the curves.
}
