% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exp_vec}
\alias{exp_vec}
\title{Computes the polynomial vector for local polynomial estimation}
\usage{
exp_vec(x, l)
}
\arguments{
\item{x}{Array, containing the evaluation points to compute over.}

\item{l}{Integer, containing the order of the polynomial.}
}
\value{
Array.
}
\description{
Computes the polynomial vector for local polynomial estimation
}
