% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{generate_curves_common}
\alias{generate_curves_common}
\title{Generate common design curves in the functional data framework.}
\usage{
generate_curves_common(
  points_vector,
  hurst,
  distortion_model = function(x) x,
  variance_fun = function(x) 1,
  N,
  sigma0 = 0.1,
  hetero = TRUE,
  novar = FALSE,
  norm_cov = FALSE,
  ...
)
}
\arguments{
\item{points_vector}{Vector containing the common sampling points for all
curves.}

\item{hurst}{Hurst function.}

\item{distortion_model}{Distortion function \code{A(.)} for the time points.}

\item{variance_fun}{Variance function.}

\item{N}{Number of curves to simulate.}

\item{sigma0}{Numeric, the baseline noise level of curves.}

\item{hetero}{Boolean, indicating whether to generate heteroscedastic curves.
See \code{sigma_het} function for more details.}

\item{novar}{Boolean, indicating whether variance function should not be
matched. \code{TRUE} results in variance not being matched.}

\item{norm_cov}{Boolean, indicating whether covariance should be normalised
by constant associated to multi-fractional brownian motion.}
}
\value{
A list, containing the ideal and sampled curves, together with other
auxiliary information.
}
\description{
Generate common design curves in the functional data framework.
}
