% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{generate_KL}
\alias{generate_KL}
\title{Generates functional data using Karhunen-Loève decomposition}
\usage{
generate_KL(
  lambda,
  psi,
  K,
  points_list,
  mu,
  sigma0,
  hetero,
  zeta_fam = c("gaussian", "unif", "chi2")
)
}
\arguments{
\item{lambda}{Vector of eigenvalues.}

\item{psi}{List, containing the following elements:
\itemize{
\item \strong{t} Vector of evaluation points for the eigenfunctions.
\item \strong{x} Matrix, with each column containing the j-th eigenfunction
at evaluation points \code{t}.
}}

\item{K}{Numeric, the number of Fourier bases used to approximate
the eigenfunctions. K should be greater than the number of input
eigenfunctions you are trying to approximate.}

\item{points_list}{List, containing the evaluation points for each
curve.}

\item{mu}{List, containing the following elements:
\itemize{
\item \strong{t} Vector of evaluation points for the mean function.
\item \strong{x} Vector containing the observed points of the mean function.
}}

\item{sigma_sd}{Vector / Numeric, containing the standard deviation
structure of the gaussian noise to be added to sample paths.}

\item{xi_fam}{Character, indicating the distribution to generate
the scores from. Gaussian, uniform and chi-squared distributions
are currently available.}
}
\value{
List, with each element containing the following:
-\strong{curves} List of curves, containing a vector of evaluation points
and observed points.
-\strong{mu} List, containing a vector of evaluation points and
observed points. The observed points are computed based on the closest point
from the mean function on the input evaluation points and the evaluation point
of the curves.
-\strong{zeta} Vector of length K, containing the true scores.
-\strong{v} Matrix with the columns containing the approximated jth eigenfunction.
-\strong{sigma} Vector containing the gaussian noise added each curve.
}
\description{
Sample paths of a stochastic process are generated using the
Karhunen-Loève decomposition, for given eigenvalues and
eigenfunction pairs. The input eigenfunctions are approximated
by orthogonal Fourier bases using a LASSO regression, before a
gram-schmidt orthonormalisation is performed so that the
approximated eigenfunctions remain orthonormal.
}
