% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{hurst_logistic}
\alias{hurst_logistic}
\title{Generate Hurst function with logistic change points}
\usage{
hurst_logistic(
  t_vec,
  h_left = 0.5,
  h_right = 0.8,
  slope = 15,
  change_point_position = 0.5
)
}
\arguments{
\item{t_vec}{Vector of sampling points.}

\item{h_left}{First Hurst Index.}

\item{h_right}{Second Hurst Index}

\item{slope}{Slope of transition between \code{h_left} and \code{h_right}.}

\item{change_point_position}{Change point position between two Hurst indexes.}
}
\value{
A vector of length \code{t_vec} containing the Hurst function.
}
\description{
Generate logistic Hurst functions with two Hurst indexes.
}
