% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{hurst_piecewise}
\alias{hurst_piecewise}
\title{Generate piecewise Hurst function.}
\usage{
hurst_piecewise(t, h_left = 0.2, h_right = 0.8, change_point_position = 0.5)
}
\arguments{
\item{h_left}{First Hurst Index.}

\item{h_right}{Second Hurst Index}

\item{change_point_position}{Change point position between two Hurst indexes.}

\item{t_vec}{Vector of sampling points.}
}
\value{
A vector of length \code{t_vec} containing the Hurst function.
}
\description{
Generate piecewise Hurst functions with two Hurst indexes. Will have
discontinuous jumps, use with caution.
}
