% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate1D}
\alias{interpolate1D}
\title{Perform one dimensional interpolation}
\usage{
interpolate1D(x, y, xout, type = "linear")
}
\arguments{
\item{x}{Vector, providing the sampling points to be interpolated from.}

\item{y}{Vector, providing the observed points to be interpolated from.}

\item{xout}{Vector, providing where interpolation should take place.}

\item{type}{String, specifying the interpolation method. Choices are either
"linear" or "constant".}
}
\value{
List, with the following components
\itemize{
\item \strong{x} Vector, containing the interpolated sampling points.
\item \strong{y} Vector, containing the interpolated points.
}
}
\description{
Interpolation using \code{approx} from \code{stats}, where the extrapolation is done
by using the closest observed point.
}
\references{
Wang S., Patilea V., Klutchnikoff N. (2023+) - Adaptive Functional
Principal Components Analysis
}
