% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate2D}
\alias{interpolate2D}
\title{Perform two dimensional interpolation}
\usage{
interpolate2D(x, y, z, xout, yout)
}
\arguments{
\item{x}{Vector, providing the \code{x} coordinates to be interpolated from.}

\item{y}{Vector, providing the \code{y} coordinates to be interpolated from.}

\item{z}{Matrix, containing the \code{z(x, y)} values to be interpolated from.}

\item{xout}{Vector, providing the \code{x} coordinates where interpolation should
take place.}

\item{yout}{Vector, providing the \code{y} coordinates where interpolation should
take place.}
}
\value{
Matrix, containing the interpolated points for every \code{z(xout, yout)}.
}
\description{
Bilinear interpolation using \code{intp} package.
}
\references{
Wang S., Patilea V., Klutchnikoff N. (2023+) - Adaptive Functional
Principal Components Analysis
}
