% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intp_list}
\alias{intp_list}
\title{Interpolates elements of a list}
\usage{
intp_list(list_in, xout)
}
\arguments{
\item{list_in}{List, containing
-\strong{t} Vector of evaluation points in which the elements of the list were
computed on.
-Vector or matrices of to be interpolated.}

\item{xout}{Vector, containing the desired output evaluation points.}
}
\value{
List, with each element interpolated onto the grid of \code{xout}.
}
\description{
Performs 1D and 2D interpolation of elements in a list
}
