% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mean.R
\name{learn_mean}
\alias{learn_mean}
\title{Estimates smoothed mean from a dataset}
\usage{
learn_mean(df, k = 50)
}
\arguments{
\item{df}{Dataframe, with rows indexing the curves and columns indexing
the time points.}

\item{k}{Number of basis functions to be used in fit}
}
\value{
A \code{glmnet} object, which can be used as input to \code{predict_mean} to
evaluate the smoothed mean on a specified grid.
}
\description{
Smooths the empirical mean of a dataset using a lasso penalty.
}
