% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_recursive.R
\name{locpoly_smooth}
\alias{locpoly_smooth}
\title{Performs presmoothing for the purpose of parameter estimation}
\usage{
locpoly_smooth(curves, grid)
}
\arguments{
\item{curves}{A list containing the raw data points with two entries:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Observed Points
}}

\item{grid}{Grid of points to estimate parameters. Must be a matrix, with
the rows representing the 3 time points, \verb{t1, t2, t3}, and the columns
are the points on the estimation grid.}
}
\value{
An \verb{G x N x 3} array, where \code{G} is the number of points on the
the grid and \code{N} is the number of curves.
}
\description{
\code{locpoly_smooth} epresmooths raw curves, for the purpose of
parameter estimation. Cross-validation is first done to determine
a local bandwidth, before a correction is performed based on thereotical
grounding.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
