% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_optimised.R
\name{lseq}
\alias{lseq}
\title{Generate exponential grid for optimisation}
\usage{
lseq(from = 1, to = 100, length.out = 51)
}
\arguments{
\item{from}{Starting point of grid.}

\item{to}{End point of grid.}

\item{length.out}{Number of points in the grid.}
}
\value{
A vector containing grid values.
}
\description{
Generate exponential grid for optimisation
}
\examples{
lseq(from = 1, to = 100, length.out = 51)
}
