% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_emp}
\alias{mean_emp}
\title{Computes the empirical mean of functional data}
\usage{
mean_emp(curves_learn, curves_online)
}
\arguments{
\item{curves_learn}{List of learning curves, containing the following as elements:
\itemize{
\item \strong{$t} Vector of sampling points.
\item \strong{$x} Vector of observed points.
}}

\item{curves_online}{List of online curves, containing the following as elements:
\itemize{
\item \strong{t} Vector of sampling points.
\item \strong{$x} Vector of observed points.
}}
}
\value{
List, containing the mean function at the sampling points of the
online set and the interpolated points from the learning set.
}
\description{
Given a learning and online set of curves, the empirical mean on the
sampling points of the online set is computed. The empirical mean is first
computed on the commonly sampled points of the learning set (assumed to be
densely observed, without noise), before interpolation is performed onto
the sampling points of the online set (assumed to be on a randomly designed
grid).
}
