% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mean.R
\name{mean_ll}
\alias{mean_ll}
\title{Estimate the mean function using adaptive bandwidths}
\usage{
mean_ll(
  data,
  grid_bandwidth = lseq(0.005, 0.1, length.out = 151),
  grid_smooth = seq(0, 1, length.out = 101),
  k0 = 2
)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used
in estimation.}
}
\value{
A list, containing
\itemize{
\item \strong{$params} Tibble with estimated parameters.
\item \strong{$mu_hat} Vector containing the mean function at each point
\code{t} in \code{grid_smooth}.
}
}
\description{
\code{mean_ll} estimates the mean from curves smoothed using an
adaptive bandwidth, typically the one returned by \code{smooth_curves_mean}.
}
