% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_optimised.R
\name{mean_plugin_evalues}
\alias{mean_plugin_evalues}
\title{Computes the mean function using a plugin bandwidth for
eigenvalue estimation}
\usage{
mean_plugin_evalues(curves, smoothed_curves, bandwidth, grid_smooth, k0)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{smoothed_curves}{List containing the smoothed curves, with
each element containing the values of one curve.}

\item{bandwidth}{Matrix of bandwidths for each point \verb{(s,t)}.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used
in estimation.}
}
\value{
A list containing
\itemize{
\item \strong{$mu} Mean curve.
\item \strong{$wt} Weights \code{wi} used to compute the mean function.
}
}
\description{
\code{mean_plugin_evalues} computes the mean function used to center
the smoothed curves for the purpose of eigenvalue estimation.
}
