% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mu_ISE}
\alias{mu_ISE}
\title{Adaptive estimation of the mean function}
\usage{
mu_ISE(Y_list, xout, param_list, bw_grid)
}
\arguments{
\item{Y_list}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{xout}{Vector containing the evaluation points for smoothing curves.}

\item{param_list}{List, containing the following estimated parameters, for
example from \code{estimate_parameters_FPCA}.}

\item{bw_grid}{Vector containing the grid of bandwidths to optimise over.}
}
\value{
Vector containing the estimated mean function.
}
\description{
Estimates the mean function given a list of functional data, by adapting to
the regularity of the sample paths. Adaptation is obtained through bandwidth
selection, which is taken to minimise the mean integrated squared error (MISE)
of the mean function.
}
