% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{N_gamma}
\alias{N_gamma}
\title{Computes the variance rate term of the covariance function}
\usage{
N_gamma(curves, x, bw)
}
\arguments{
\item{curves}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{x}{Vector containing the evaluation points in one dimension. The final
quantity is computed over the cartesian product of (\code{x}, \code{x}).}

\item{bw}{Vector containing the grid of bandwidths.}
}
\value{
List, containing two elements:
\itemize{
\item \strong{WN_bi} Array of dimension G x G x H, where G is the length of \code{x} and
\code{H} is the length of \code{bw}. This term contains the effective sample size for
a given bandwidth in the estimation of the covariance function.
\item \strong{Ngamma} Array of dimension G x G x H, where G is the length of \code{x} and
\code{H} is the length of \code{bw}. This term contains the variance term in the risk
bounds of the covariance function, which contributes to the rate of convergence.
}
}
\description{
This quantity computes the term that governs that rate of convergence of
the variance term for the covariance function.
}
