% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalise_sign}
\alias{normalise_sign}
\title{Normalise the sign of estimated eigenfunctions}
\usage{
normalise_sign(efunction, efunction_true)
}
\arguments{
\item{efunction}{Matrix containing the estimated eigenfunctions.}

\item{efunction_true}{Matrix containing the true eigenfunctions.}
}
\value{
Matrix, containing the normalised eigenfunctions.
}
\description{
Since estimated eigenfunctions are only defined up to a sign, normalisation
is required to have identical signs with the true eigenfunctions for
the purposes of comparison.
}
