% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NW_max}
\alias{NW_max}
\title{Computes the maximum (over observed points) Nadaraya-Watson weights}
\usage{
NW_max(curves, x, bw)
}
\arguments{
\item{curves}{List, containing the following elements:
-\strong{t} Vector, containing the sampling points of curves.
-\strong{x} Vector, containing the observed points.}

\item{x}{Vector, containing the evaluation points.}

\item{bw}{Vector, containing the grid of bandwidth points.}
}
\value{
Array of dimension \verb{G x N x H}, corresponding to the
number of evaluation points, curves, and bandwidth points respectively.
}
\description{
Computes the maximum (over observed points) Nadaraya-Watson weights
}
