% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ortho_funmat}
\alias{ortho_funmat}
\title{Orthonormalise the eigenfunction matrix}
\usage{
ortho_funmat(X, t)
}
\arguments{
\item{X}{Matrix, containing the eigenfunctions in the columns to be
orthonormalised.}

\item{t}{Vector containing the evaluation points of the eigenfunction.}
}
\value{
Matrix, containing the orthonormalised eigenfunctions.
}
\description{
Given a matrix where the columns are composed of the linearly
independent eigenfunctions, the QR decomposition is returned. Normalisation
is taking performed to make the eigenfunctions of unit norm, and have the
same sign as the input eigenfunctions.
}
