% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{PACE_adapt}
\alias{PACE_adapt}
\title{Computes PACE estimates with learning and online set}
\usage{
PACE_adapt(Y_learn, Y_online, x_out, h_out, param_list, h_power, n_scores)
}
\arguments{
\item{Y_learn}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{Y_online}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{x_out}{Vector containing the evaluation points to smooth curves.}

\item{h_out}{Vector containing the grid of bandwidths.}

\item{h_power}{Numeric indicating the power to raise bandwidth to when
estimating sigma for the purposes of diagonal correction for the covariance function.}

\item{n_scores}{Numeric indicating the number of scores to estimate.}

\item{param_out}{Vector containing the evaluation points for parameter
estimation.}

\item{ncv_param}{Numeric containing the number of curves used for learning
the cross-validation bandwidth in presmoothing.}

\item{h_q}{Numeric containing the quantile selected from the set of
cross-validation bandwidths.}

\item{gamma_H}{Numeric, indicating the gamma to be used for estimation of H.
See \code{estimate_regularity}.}

\item{gamma_L}{Numeric, indicating the gamma to be used for estimation of L.
See \code{estimate_regularity}.}

\item{n_knots}{Numeric indicating the number of knots used for smoothing H and L,
where splines are used.}
}
\value{
Matrix of dimension K x N, where K is the number of scores and
N is the number of curves in \code{Y_online}.
}
\description{
Adaptive PACE estimates are computed given a list of learning curves and
a list of online curves. Quantities are estimated on the set of learning
curves, before being interpolated onto the sampling points given by a list
of online curves. Each quantity is adaptively and separately computed
using optimal estimates in the sense of mean-integrated squared error (MISE).
}
