% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PACE_emp}
\alias{PACE_emp}
\title{Computes PACE estimates based on a learning and online set of curves}
\usage{
PACE_emp(curves_learn, curves_online, sigma)
}
\arguments{
\item{curves_learn}{List of learning curves, containing the following as elements:
\itemize{
\item \strong{$t} Vector of sampling points.
\item \strong{$x} Vector of observed points.
}}

\item{curves_online}{List of online curves, containing the following as elements:
\itemize{
\item \strong{t} Vector of sampling points.
\item \strong{$x} Vector of observed points.
}}

\item{sigma}{Numeric, containing the noise of the online curves.}
}
\value{
List, containing the scores for each online curve.
}
\description{
Given a list of commonly sampled functional data, the conditional
expectation of the scores given the observed values are computed. The
plug-in estimates of the conditional expectation are the empirical
mean, covariance, and eigen-elements.
}
