% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{plot.curves_ideal_observed}
\alias{plot.curves_ideal_observed}
\title{Plot functions for generated true curves}
\usage{
\method{plot}{curves_ideal_observed}(curves, which_curves = 1:length(curves$curves), ...)
}
\arguments{
\item{curves}{List, containing the ideal curves.}

\item{which_curves}{Vector, indicating the curves to be plotted.}
}
\description{
Plot functions for generated true curves
}
