% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{powerconsumption}
\alias{powerconsumption}
\title{Powerconsumption data set}
\format{
A tibble with 1358 rows and 1440 columns, with rows representing
the daily curves and columns representing 1 minute samples over a 24-hour
period.
\describe{
Measurements of voltage consumption in one household with a one-minute
sampling rate in a 24 hour period, with data taken over almost 4 years. Days with missing
values were removed from the dataset.
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Individual+household+electric+power+consumption}
}
\usage{
powerconsumption
}
\description{
A subset of the powerconsumption dataset from the UCI Repository
}
\keyword{datasets}
