% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict_mean}
\alias{predict_mean}
\title{Evaluates the smoothed mean on a grid}
\usage{
predict_mean(u, model, lambda, k = 50, scale)
}
\arguments{
\item{u}{Vector of sampling points on which the mean which be computed at.}

\item{model}{\code{glmnet} object containing the mean curve.}

\item{lambda}{Numeric, representing the value of the regularisation parameter.}

\item{k}{Number of basis functions used to fit the data.}

\item{scale}{Boolean. If \code{TRUE}, scales the output by the global mean.}
}
\value{
A numeric vector containing the mean curve evaluated on the vector of
grid points \code{u}.
}
\description{
Given a \code{glmnet} object, for example from the \code{learn_mean} function,
\code{predict_mean} computes the function at specific evaluation points.
}
