% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{presmoothing_FPCA}
\alias{presmoothing_FPCA}
\title{Presmoothing with cross-validation}
\usage{
presmoothing_FPCA(data, N0, grid_smooth, q)
}
\arguments{
\item{data}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{N0}{Numeric, number of curves randomly selected for cross-validation.}

\item{grid_smooth}{Vector, grid of t2 points used to presmooth curves.}

\item{q}{Numeric, quantile of cv bandwidth used to presmooth curves.}
}
\value{
A list, containing two elements:
\itemize{
\item \strong{$presmoothed_curves} List of presmoothed curves.
\item \strong{$cv_bandwidth} Numeric, containing the bandwidth used.
}
}
\description{
Performs presmoothing of functional data, using a Nadaraya-Watson smoother
and a bandwidth learned from cross-validation (cv) on a subset of curves.
Bandwidth used is a q-th quantile of the set of those learned from cv on
each curve. Selected curves are uniformly sampled.
}
\references{
Wang S., Patilea V., Klutchnikoff N., (2023+) - Adaptive
Functional Principal Components Analysis
}
