% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{presmoothing_splines}
\alias{presmoothing_splines}
\title{Estimates local regularity using splines}
\usage{
presmoothing_splines(data, t0_list = seq(0.1, 0.9, l = 20))
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{t0_list}{Vector of sampling points which presmoothing
is performed.}
}
\value{
A vector containing \code{H(t)} evaluated at \code{t0_list}.
}
\description{
Estimates the local regularity \code{H(t)} by presmoothing, where
presmoothing is performed using splines on irregularly sampled curves.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
