% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{psi_approx}
\alias{psi_approx}
\title{Approximate input functions using Fourier bases}
\usage{
psi_approx(psi_t, psi_x, K)
}
\arguments{
\item{psi_t}{Vector containing the evaluation points.}

\item{psi_x}{Vector containing the observed points.}

\item{K}{Numeric indicating the number of Fourier bases functions to use
for approximating input function.}
}
\value{
Vector containing the coefficients from LASSO regression.
}
\description{
Performs approximation of input functions (primarily used for eigenfunctions
) using orthonormal Fourier bases and LASSO regression.
}
