% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{sigma_het}
\alias{sigma_het}
\title{Generates gaussian noise for function data}
\usage{
sigma_het(sigma0, hetero = TRUE, t_vec)
}
\arguments{
\item{sigma0}{Numeric, baseline noise level where the noise will be centered
around.}

\item{hetero}{Boolean, where TRUE indicate}

\item{t_vec}{Vector, containing the sampling points for}
}
\value{
A list containing the noise for each curve if \code{hetero = TRUE}. If
not, a numeric will be returned.
}
\description{
Generates possibly heteroscedastic gaussian noise of functional data,
based on the observed time points of one curve.
Heteroscedasticity is generated using sin functions.
}
