% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{smooth_curves}
\alias{smooth_curves}
\title{Smooth curves with a given bandwidth}
\usage{
smooth_curves(data, grid, bandwidth, drv = 0, deg = drv + 1)
}
\arguments{
\item{data}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{grid}{Vector / List containing the grid of points to smooth curves. If
the input is a list, then smoothing is performed matching the i-th curve
in \code{data} and the i-th element on the grid. Currently available only for the
Nadaraya-Watson smoother.}

\item{bandwidth}{Numeric containing the bandwidth.}

\item{drv}{Numeric, the order of derivative to be estimated.}

\item{deg}{Numeric, the order of polynomial fit to be used.}
}
\value{
List, containing the smoothed curves.
}
\description{
Performs smoothing of curves for a given bandwidth using the either the local
linear or Nadaraya-Watson smoother.
}
