% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eigenfunctions.R
\name{smooth_curves_efunctions}
\alias{smooth_curves_efunctions}
\title{Smooth curves with adaptive eigenfunction bandwidth}
\usage{
smooth_curves_efunctions(
  curves,
  grid_bandwidth,
  grid_smooth,
  k0,
  nfunctions,
  params
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used in
estimation.}

\item{nfunctions}{The number of eigenfunctions to be kept.}

\item{params}{A tibble, containing the following parameters:
\itemize{
\item \strong{$t$} Sampling points.
\item \strong{$H} Estimated Hölder exponents.
\item \strong{$L} Estimated Hölder constants.
\item \strong{$sigma} Estimated noise.
\item \strong{$mu} Estimated density of time points.
estimates it.
}}
}
\value{
A list of two elements, containing
\itemize{
\item A list of curves, with
\itemize{
\item $t Sampling points.
\item $x Observed points.
}
\item A vector of bandwidths used to smooth each curve.
}
}
\description{
\code{smooth_curves_efunctions} smooths irregularly sampled curves onto a specified
grid to approximate densely observed curves in functional data. Used
primarily to estimate eigenfunctions downstream.
}
