% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eigenvalues.R
\name{smooth_curves_evalues}
\alias{smooth_curves_evalues}
\title{Smooth curves with adaptive eigenvalue bandwidth}
\usage{
smooth_curves_evalues(curves, grid_bandwidth, grid_smooth, k0, nvalues, params)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used in
estimation.}

\item{nvalues}{The number of eigenvalues to be kept.}

\item{grid_param}{Grid of points to estimate parameters.}

\item{sigma}{Noise level of curves.}

\item{mu0}{Density lower bound for time points.}
}
\value{
A list of two elements, containing
\itemize{
\item A list of curves, with
\itemize{
\item $t Sampling points.
\item $x Observed points.
}
\item A vector of bandwidths used to smooth each curve.
}
}
\description{
\code{smooth_curves_evalues} smooths irregularly sampled curves onto a specified
grid to approximate densely observed curves in functional data. Used
primarily to estimate eigenvalues downstream.
}
\examples{
smooth_curves_evalues(curves = curves_list,
grid_bandwidth = seq(0, 1, length.out = 151),
grid_smooth = seq(0, 1, length.out = 101),
k0 = 1,
grid_param = seq(0, 1, length.out = 20),
sigma = 0.1, mu0 = 1, nvalues = 10)
}
