% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_optimised.R
\name{smooth_curves_mean}
\alias{smooth_curves_mean}
\title{Smooth curves using adaptive bandwidth for mean function estimation}
\usage{
smooth_curves_mean(curves, params, grid_bandwidth, grid_smooth, k0)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{params}{Tibble of parameters, containing
\itemize{
\item \strong{$H} Estimated regularity.
\item \strong{$L} Estimated Hölder constant.
\item \strong{$sigma} Noise level of curves.
\item \strong{$mu0} Density lower bound for time points. Defaults to NULL,
which estimates it.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used
in estimation.}
}
\value{
A list, containing
\itemize{
\item \strong{$params} Tibble with estimated parameters.
\item \strong{$x_hat} Matrix of curves, with columns indexing each curve and
rows indexing each point on \code{grid_smooth}.
}
}
\description{
\code{smooth_curves_mean} smooths each curve using an input bandwidth,
typically the one returned by \code{estimate_bandwidth_mean}. Smoothing is
targeted for the purpose of mean function estimation.
}
