% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_optimised.R
\name{smooth_curves_mean_plugin_cov}
\alias{smooth_curves_mean_plugin_cov}
\title{Smooth curves using a plugin bandwidth for covariance estimation}
\usage{
smooth_curves_mean_plugin_cov(curves, grid_smooth, k0, bandwidth)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used
in estimation.}

\item{bandwidth}{Matrix of bandwidths for each point \verb{(s,t)}.}
}
\value{
A list containing smoothed curves,
}
\description{
\code{smooth_curves_mean_plugin_cov} smooths each curve using a plugin
bandwidth, typically to estimate the mean used to center the
estimated covariance function. Used as an auxiliary function
in \code{covariance_ll}.
}
