% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{bertin_smoother}
\alias{bertin_smoother}
\title{Smooths curves using modified Nadaraya-Watson smoother}
\usage{
bertin_smoother(data, grid, bandwidth, beta)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid}{List, consisting of the following elements:
\itemize{
\item $x Sorted grid on which to smooth curves.
\item $ix Indexes of the sorted grid build from the original vector of grid
points.
}}

\item{bandwidth}{Bandwidth used to smooth curves, usually
calculated from \code{bertin_bandwidth}.}

\item{beta}{Initialised regularity.}
}
\value{
An 3D array of dimension G x N x 3, where G is the number of
\code{t2} points, N is the number of curves, and the 3rd dimension is for each
of the \code{t1}, \code{t2} and \code{t3} time points respectively.
}
\description{
\code{bertin_smoother} performs curve smoothing using a kernel, referred to as
the Bertin kernel (see \code{bertin_kernel}). Usually used as an auxiliary
function for others such as \code{presmoothing}.
}
\references{
Bertin L, (2004) - Minimax exact constant in sup-norm for
nonparametric regression with random design.
}
