% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPCA.R
\name{bw_FPCA}
\alias{bw_FPCA}
\title{Compute FPCA bandwidth without index specific constants}
\usage{
bw_FPCA(
  data,
  parameters,
  h_grid,
  t_grid,
  inflate = TRUE,
  interp_type = "linear"
)
}
\arguments{
\item{data}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{parameters}{List of parameters, which can be obtained from
\code{estimate_parameters_FPCA}.}

\item{h_grid}{Vector of points for bandwidth optimization.}

\item{t_grid}{Vector of sampling points to smooth curves.}

\item{inflate}{Boolean, indicating whether to inflate bandwidth as a
correction for the discretization error from estimating regularity.}
}
\value{
Numeric, containing the optimal bandwidth.
}
\description{
Performs estimation of the FPCA bandwidth when all the \emph{index specific}
constants are dropped for the risk bound, resulting in a single bandwidth
for both the eigenvalues and the eigenfunctions, instead of a different
bandwidth for each eigenvalue and eigenfunction.
}
\references{
Wang S., Patilea V., Klutchnikoff N., (2023+) - Adaptive
Functional Principal Components Analysis
}
