% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_covariance_norm.R
\name{estimate_bandwidth_covariance_norm}
\alias{estimate_bandwidth_covariance_norm}
\title{Estimate global adaptive bandwidth for covariance function}
\usage{
estimate_bandwidth_covariance_norm(
  curves,
  grid_bandwidth,
  grid_smooth,
  k0,
  grid_tibble
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used in
estimation.}

\item{grid_param}{Grid of points to estimate parameters.}

\item{sigma}{Noise level, if known. Defaults to NULL, which estimates it.}

\item{mu0}{Density lower bound for time points. Defaults to NULL, which
estimates it.}
}
\value{
A list with two elements, containing
\itemize{
\item A global bandwidth.
\item A list, containing
\itemize{
\item \strong{$moments} Estimated moments.
\item \strong{$constants} Estimated Hölder constants \code{H(t)} and \code{L(t)}.
\item \strong{$kernel_int} Integrated constant of kernel.
\item \strong{$N_gamma_ts} Harmonic mean of the number of points used for
smoothing each curve at \code{(t|s)}.
\item \strong{$N_gamma_st} Harmonic mean of the number of points used for
smoothing each curve at \code{(s|t)}.
\item \strong{$WN} Total number of curves selected for estimation.
}
}
}
\description{
\code{estimate_bandwidth_covariance_norm} estimates the global adaptive bandwidth
in terms of integrated squared risk, as opposed to
\code{estimate_bandwidth_covariance}, which estimates a pointwise bandwidth.
}
