% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_recursive.R
\name{estimate_density}
\alias{estimate_density}
\title{Estimate minimum density of sample points}
\usage{
estimate_density(data)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}
}
\value{
A number.
}
\description{
\code{estimate_density} estimates the minimum density of time points using
the kernel density estimator.
}
