% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_recursive.R
\name{estimate_H}
\alias{estimate_H}
\title{Estimated the local regularity based on presmoothing}
\usage{
estimate_H(smoothed_curves, weights)
}
\arguments{
\item{smoothed_curves}{An array with dimensions \verb{G x N x 3} or \verb{G x N x 4},
where the rows are the number of points on the grid used for estimation,
the columns are the curves, the 3rd dimension depends on whether different
curves are used for estimation.
Should use either \code{locpoly_smooth} or \code{cv_smooth} as an auxiliary function
to obtain the proper inputs.}
}
\value{
A vector, whose length is equal to the number of sampling
points the presmoothed curves takes values in.
}
\description{
Estimates the local regularity with supplied presmoothed curves.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
