% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_H0}
\alias{estimate_H0}
\title{Performs estimation of Hölder exponent}
\usage{
estimate_H0(presmoothed_data)
}
\arguments{
\item{presmoothed_data}{A list, containing
\itemize{
\item \strong{$x} Smoothed data points.
}}
}
\value{
A vector containing the estimated values at the
sampling points of presmoothed curves.
}
\description{
\code{estimate_H0} estimates the Hölder constant of presmoothed curves,
typically as output from the function \code{presmoothing}.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
