% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_recursive.R
\name{estimate_holder_quantities}
\alias{estimate_holder_quantities}
\title{Performs estimation of parameters}
\usage{
estimate_holder_quantities(curves, grid_param)
}
\arguments{
\item{curves}{A list containing the raw data points with two entries:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Observed Points
}}

\item{grid_param}{Grid of points to estimate parameters.}
}
\value{
A list with two elements:
\itemize{
\item \strong{$params} Tibble of parameters.
\item \strong{$smoothed_curves} Smoothed_curves.
}
}
\description{
\code{estimate_holder_quantities} estimates the parameters used for
downstream analysis, typically for functions such as
\code{evalues_adaptive}. Cross validation to determine the local
bandwidth for presmoothing, before a correction is performed
based on theoretical developments.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
