% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_L0}
\alias{estimate_L0}
\title{Performs estimation of Hölder constant}
\usage{
estimate_L0(presmoothed_data, H0_list, t0_list, M)
}
\arguments{
\item{presmoothed_data}{A list, containing
\itemize{
\item \strong{$x} Smoothed data points.
}}

\item{H0_list}{Vector containing Hölder exponents, typically as
output from \code{estimate_H0}.}

\item{t0_list}{Vector containing the sampling points of \code{H0_list}.}

\item{M}{Mean number of points on each curve.}
}
\value{
A vector containing the estimated values at the
sampling points of presmoothed curves.
}
\description{
\code{estimate_L0} estimates the Hölder constant of presmoothed curves,
typically as output from the function \code{presmoothing}. Smooths the
input values of \code{H0} by splines to get more stable results.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
