% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{estimate_parameters_FPCA}
\alias{estimate_parameters_FPCA}
\title{Estimate parameters for FPCA}
\usage{
estimate_parameters_FPCA(data, grid_points, gamma_H, gamma_L, nknots)
}
\arguments{
\item{data}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{grid_points}{Vector containing the sampling points to estimate the
parameters.}

\item{gamma_H}{Numeric, power to be used in computing delta for Hurst index.}

\item{gamma_L}{Numeric, power to be used in computing delta for Holder constant.}

\item{nknots}{Numeric, number of knots used for smoothing H and L.}
}
\value{
List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points where estimation is performed.
\item \strong{$H} Vector of Hurst indexes.
\item \strong{$L} Vector of Hölder constants.
\item \strong{$sigma} Vector of conditional variance / noise.
\item \strong{$density} Numeric containing minimum density of sampling points.
\item \strong{$moment2} Vector of second moments of \code{Xt}.
\item \strong{$moment2prod} Matrix of second moment of products \code{XtXs}.
}
}
\description{
Performs estimate of parameters used in FPCA, in particular the local
regularity, moments, noise, and density of sampling points. See
separate internal functions for more details.
}
\references{
Wang S., Patilea V., Klutchnikoff N., (2023+) - Adaptive
Functional Principal Components Analysis
}
