% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{Estimate conditional variance}
\usage{
estimate_sigma(data, sigma_grid, h, h_power = 0.9)
}
\arguments{
\item{data}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{sigma_grid}{Vector containing the sampling points to estimate
the conditional variance.}

\item{h}{Optional numeric containing the bandwidth h, which is only
used for correcting the diagonal band.}

\item{h_power}{Numeric, power to raise the bandwidth to when using
sigma for diagonal correction for the covariance function.}
}
\value{
Vector containing sigma.
}
\description{
Estimates the conditional variance of curves using ordered statistics,
correcting for sparse regimes by selecting only points that are sufficiently
close enough to each other.
}
\references{
Wang S., Patilea V., Klutchnikoff N., (2023+) - Adaptive
Functional Principal Components Analysis
}
