% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_recursive.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{Estimate pointwise noise level of curves using only observed points}
\usage{
estimate_sigma(data, grid_param = seq(0.1, 0.9, length.out = 20))
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_param}{Vector containing the sampling points to estimate
the noise.}
}
\value{
Vector with the same length as \code{grid_param}.
}
\description{
\code{estimate_sigma} estimates the variance of curves, using only
information from curves.
}
