% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues.R
\name{evalues_adaptive}
\alias{evalues_adaptive}
\title{Estimate eigenvalues of functional data with adaptive bandwidths}
\usage{
evalues_adaptive(
  curves,
  grid_bandwidth,
  grid_smooth,
  k0,
  grid_param,
  sigma = NULL,
  mu0 = NULL,
  nvalues = 10
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used in
estimation.}

\item{grid_param}{Grid of points to estimate parameters.}

\item{sigma}{Noise level, if known. Defaults to NULL, which estimates it.}

\item{mu0}{Density lower bound for time points. Defaults to NULL, which
estimates it.}

\item{nvalues}{The number of eigenvalues to be kept.}
}
\value{
A list with two elements, containing
\itemize{
\item Eigenvalues.
\item Bandwidths used for smoothing.
}
}
\description{
\code{evalues_adaptive} estimates eigenvalues using an adaptive bandwidth.
Curves can be irregularly sampled, possess heterogenous
degrees of smoothness and be relatively sparse (but should have
at least 20 points per curve). Uses the "smooth first, then estimate"
approach.
}
\examples{
evalues_adaptive(curves = curves_list,
grid_bandwidth = seq(0, 1, length.out = 151),
grid_smooth = seq(0, 1, length.out = 101),
k0 = 1,
grid_param = seq(0, 1, length.out = 20),
sigma = 0.1, mu0 = 1, nvalues = 10)
}
\references{
Patilea V., Wang S. (2022+) - Adaptive Functional
Principal Components Analysis
}
