% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{generate_curves}
\alias{generate_curves}
\title{Generate curves in the functional data framework.}
\usage{
generate_curves(
  points_list,
  hurst,
  sigma = 0.1,
  add_regular_grid_of_size = 0,
  tau,
  L = 1
)
}
\arguments{
\item{points_list}{List where each element contains the sampling points
for one curve.}

\item{hurst}{Hurst function.}

\item{sigma}{Noise level of curves.}

\item{distortion_model}{Distortion function \code{A(.)} for the time points.}

\item{variance_fun}{Variance function \code{v(t)^2}.}

\item{regular_grid}{Vector, containing the grid on which the true curves
will lie.}

\item{add_one_to_hurst}{Boolean, for use if fractional regularity (1 + ...)
is desired.}

\item{norm_cov}{Boolean, indicating whether covariance should be normalised
by constant associated to multi-fractional brownian motion.}
}
\value{
A list, containing the ideal and sampled curves, together with other
auxiliary information.
}
\description{
Generate curves in the functional data framework.
}
