% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{generate_curves}
\alias{generate_curves}
\title{Generate curves in the functional data framework.}
\usage{
generate_curves(
  points_list,
  hurst,
  distortion_model = function(x) x,
  variance_fun = function(x) 1,
  sigma0 = 0.1,
  hetero = TRUE,
  regular_grid = seq(0, 1, l = 101),
  add_one_to_hurst = FALSE,
  norm_cov = FALSE,
  novar = FALSE,
  ...
)
}
\arguments{
\item{points_list}{List where each element contains the sampling points
for one curve.}

\item{hurst}{Hurst function.}

\item{distortion_model}{Distortion function \code{A(.)} for the time points.}

\item{variance_fun}{Variance function \code{v(t)^2}.}

\item{sigma0}{Numeric, the baseline noise level of curves.}

\item{hetero}{Boolean, indicating whether to generate heteroscedastic curves.
See \code{sigma_het} function for more details.}

\item{regular_grid}{Vector, containing the grid on which the true curves
will lie.}

\item{add_one_to_hurst}{Boolean, for use if fractional regularity (1 + ...)
is desired.}

\item{norm_cov}{Boolean, indicating whether covariance should be normalised
by constant associated to multi-fractional brownian motion.}
}
\value{
A list, containing the ideal and sampled curves, together with other
auxiliary information.
}
\description{
Generate curves in the functional data framework.
}
