% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{generate_points}
\alias{generate_points}
\title{Generate random time points for functional data.}
\usage{
generate_points(N, m, distribution = runif, ...)
}
\arguments{
\item{N}{Number of curves.}

\item{m}{Average number of observed points per curve.}

\item{distribution}{The distribution of the sampling points. Defaults to the
uniform distribution.}
}
\value{
A list containing the generated time points for \code{N} curves.
}
\description{
Generates sampling points in the random design framework, where the exact
number of points per curve are generated according to a Poisson distribution
with \code{m} average number of points.
}
