% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{hurst_linear}
\alias{hurst_linear}
\title{Generate Hurst function with linear change points}
\usage{
hurst_linear(t_vec, h_left = 0.2, h_right = 0.8)
}
\arguments{
\item{t_vec}{Vector of sampling points.}

\item{h_left}{First Hurst Index.}

\item{h_right}{Second Hurst Index}
}
\value{
A vector of length \code{t_vec} containing the Hurst function.
}
\description{
Generate linear Hurst functions with two Hurst indexes.
}
