% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{presmoothing}
\alias{presmoothing}
\title{Performs presmoothing of curves}
\usage{
presmoothing(
  data,
  t0_list = seq(0.1, 0.9, l = 20),
  init_b = 1,
  init_L = 1,
  sigma = NULL,
  mu0 = NULL
)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{t0_list}{Vector of sampling points which presmoothing
is performed.}

\item{init_b}{Initialised Hölder exponent.}

\item{init_L}{Initialised Hölder constant.}

\item{sigma}{Noise level, if known. Defaults to NULL, which estimates it.}

\item{mu0}{Density lower bound for time points. Defaults to NULL,
which estimates it.}
}
\value{
A list, with the following elements:
\itemize{
\item \strong{$presmoothed}: An 3D array of dimension G x N x 3, where G is the number of
\code{t2} points, N is the number of curves, and the 3rd dimension is for each
of the \code{t1}, \code{t2} and \code{t3} time points respectively.
\item \strong{$t_list}: Matrix of time points, with the rows being the \code{t1}, \code{t2}, \code{t3}
points nad the columns are points on the estimation grid.
}
}
\description{
\code{presmoothing} performs presmoothing on irregularly sampled curves,
for the purpose of estimating parameters such as Hölder constants.
Performed using a modified Nadaraya-Watson estimator, with bandwidth
detailed in the references. A lower and upper bound on the bandwidths
are imposed in order to avoid degenerate cases.
}
\references{
Bertin L, (2004) - Minimax exact constant in sup-norm for
nonparametric regression with random design.
}
