% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_and_L_functions.R
\name{presmoothing}
\alias{presmoothing}
\title{Performs presmoothing of curves}
\usage{
presmoothing(
  data,
  t0_list = seq(0.2, 0.8, l = 20),
  init_b = 1,
  init_L = 1,
  sigma = NULL,
  mu0 = NULL
)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{t0_list}{Vector of sampling points which presmoothing
is performed.}

\item{init_b}{Initialised Hölder exponent.}

\item{init_L}{Initialised Hölder constant.}

\item{sigma}{Noise level, if known. Defaults to NULL, which estimates it.}

\item{mu0}{Density lower bound for time points. Defaults to NULL,
which estimates it.}
}
\value{
A list, containing
\itemize{
\item \strong{$t_list} Sample points used for presmoothing.
\item \strong{$x} Smoothed data points.
}
}
\description{
\code{presmoothing} performs presmoothing on irregularly sampled curves,
for the purpose of estimating parameters such as Hölder constants.
Performed using the Nadaraya-Watson estimator, with bandwidth
detailed in the references.
}
\references{
Bertin L, (2004) - Minimax exact constant in sup-norm for
nonparametric regression with random design.
}
