% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{presmoothing}
\alias{presmoothing}
\title{Presmoothing with local polynomial estimator}
\usage{
presmoothing(data, drv, xout, h)
}
\arguments{
\item{data}{List of curves, with each element/curve containing two entries:
\itemize{
\item \strong{$t} Vector of time points along each curve.
\item \strong{$x} Vector of observed points along each curve.
}}

\item{drv}{Numeric, the order of derivative to be estimated.}

\item{xout}{Vector, the evaluation points of presmoothed curves.}
}
\value{
List, containing the presmoothed curves.
}
\description{
Performs presmoothing of functional data using a local polynomial estimator.
A simple bandwidth rule of m / 3 is used, where m is the average number of
sampling poins along the curves.
}
