% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_covariance_norm.R
\name{smooth_curves_covariance_norm}
\alias{smooth_curves_covariance_norm}
\title{Smooth curves using adaptive global bandwidths for covariance function
estimation (in terms of \code{L2} norm).}
\usage{
smooth_curves_covariance_norm(
  curves,
  grid_bandwidth,
  grid_smooth,
  k0,
  param_tibble
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve
must be a list with two entries:
\itemize{
\item $t Sampling points.
\item $x Observed points.
}}

\item{grid_bandwidth}{Grid of points to estimate bandwidth.}

\item{grid_smooth}{Grid of points to smooth curves.}

\item{k0}{Minimum number of points curves must possess to be used in
estimation.}

\item{param_tibble}{A tibble, containing the following parameters:
\itemize{
\item \strong{$t$} Sampling points.
\item \strong{$H} Estimated Hölder exponents.
\item \strong{$L} Estimated Hölder constants.
\item \strong{$sigma} Estimated noise.
\item \strong{$mu0} Estimated density of time points.
}}
}
\value{
A list with two elements, containing
\itemize{
\item \strong{$smoothed_product} List containing the smoothed product \code{XtXs}.
\item \strong{$smoothed_curves} List containing the smoothed curves \code{Xt}.
\item \strong{$bandwidth} Global bandwidth used to smooth each curve.
\item \strong{$params} A list, containing
\itemize{
\item \strong{$moments} Estimated moments.
\item \strong{$constants} Estimated Hölder constants \code{H(t)} and \code{L(t)}.
\item \strong{$kernel-int} Integrated constant of kernel.
\item \strong{$N_gamma_ts} Harmonic mean of the number of points used for
smoothing each curve at \code{(t|s)}.
\item \strong{$N_gamma_st} Harmonic mean of the number of points used for
smoothing each curve at \code{(s|t)}.
\item \strong{$WN} Total number of curves selected for estimation.
}
\item \strong{$nw_weights} Nadaraya-Watson weights at optimal bandwidth value.
}
}
\description{
\code{smooth_curves_covariance_norm} smooths each curve using the bandwidth
estimated from \code{estimate_bandwidth_covariance_norm}, for the purpose
of covariance function estimation.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
